fix: split component and change layout spacing
